feat: Latest TypeScript refactoring
