+++ packages/feathers/_src/hooks/base.ts
