+++ packages/feathers/_src/version.ts
@@ -1,0 +1,1 @@
+export default 'development';
