+++ packages/feathers/_test/hooks/async.test.ts
