+++ packages/feathers/_test/hooks/before.test.ts
