+++ packages/feathers/_test/hooks/error.test.ts
