+++ packages/feathers/_test/hooks/finally.test.ts
