+++ packages/feathers/src/hooks/legacy.ts
@@ -1,0 +1,4 @@
+// export function legacyHookMixin (
+//   this: Application, service: Service<any>, path: string, options: ServiceOptions<any>
+// ) {
+// }
